Rename from Description to PopOver.
