+++ src/pages/recruiting.tsx
@@ -11,7 +11,7 @@ import RecruitingBG from '../resources/recruiting-bg.png'
 import Bg from '../components/Bg'
 import ActionCard from '../components/ActionCard'
 import PricingTable from '../components/PricingTable'
-import Description from '../components/Description'
+import PopOver from '../components/PopOver'
 import Circle from '../components/Circle'
 import Layer from '../resources/layer.svg'
 import Github from '../resources/octicons-mark-github.svg'
@@ -104,7 +104,7 @@ const RecrutingPage: React.SFC<{}> = () => (
                     <td>On-prem or private cloud</td>
                 </tr>
                 <tr>
-                    <th>SLA <Description description="Custom Service Level Agreement"/></th>
+                    <th>SLA <PopOver description="Custom Service Level Agreement"/></th>
                     <td>optional</td>
                     <td>optional</td>
                     <td></td>
@@ -116,7 +116,7 @@ const RecrutingPage: React.SFC<{}> = () => (
                     <td><Circle /></td>
                 </tr>
                 <tr>
-                    <th><strong>White Labeling</strong> <Description description="Possibility to rebrand"/></th>
+                    <th><strong>White Labeling</strong> <PopOver description="Possibility to rebrand"/></th>
                     <td></td>
                     <td></td>
                     <td></td>
